refactor: move dialog.tsx to shadcn-ui
